+++ week2/sieve.cpp
@@ -1,20 +1,20 @@
 #include <iostream>
 
-void sieve(int a, int b)
+void sieve(long long a, long long b)
 {
-  int i, j, primes[100] = {1};
+  long long i, j, primes[100] = {1};
   for(i = a; i * i < b; i++)
     if (primes[i])
       for(j = i * i; j < b; j += i) primes[j] = 0;
 
-      for(int i = a; i < b; i++){
+      for(long long i = a; i < b; i++){
           if(primes[i] == 1) 
               std::cout << i << std::endl;
       }
 }
 
 int main(){
-    int a, b;
+    long long a, b;
     std::cin >> a >> b;
     sieve(a, b);
 }
